+++ src/backend.ts
@@ -284,6 +284,10 @@ export function GetTags(params: { tags: string[] }): Promise<Tag[]> {
     return invoke("get_tags", params)
 }
 
+export function GetTagsWithoutConflict(params: { tags: string[] }): Promise<Tag[]> {
+    return invoke("get_tags_without_conflict", params)
+}
+
 export function GetItems(params: { items: ItemId[] }): Promise<Item[]> {
     return invoke("get_items", params)
 }